reorg code
