+++ evm/src/main/kotlin/org/apache/tuweni/evm/EthereumVirtualMachine.kt
@@ -456,3 +456,117 @@ interface EvmVm {
   suspend fun execute(hostContext: HostContext, fork: HardFork, msg: EVMMessage, code: Bytes): EVMResult
   fun capabilities(): Int
 }
+
+val opcodes = mapOf<Byte, String>(
+  Pair(0x00, "stop"),
+  Pair(0x01, "add"),
+  Pair(0x02, "mul"),
+  Pair(0x03, "sub"),
+  Pair(0x04, "div"),
+  Pair(0x05, "sdiv"),
+  Pair(0x06, "mod"),
+  Pair(0x06, "mod"),
+  Pair(0x07, "smod"),
+  Pair(0x08, "addmod"),
+  Pair(0x09, "mulmod"),
+  Pair(0x0a, "exp"),
+  Pair(0x0b, "signextend"),
+  Pair(0x10, "lt"),
+  Pair(0x11, "gt"),
+  Pair(0x12, "slt"),
+  Pair(0x13, "sgt"),
+  Pair(0x14, "eq"),
+  Pair(0x15, "isZero"),
+  Pair(0x16, "and"),
+  Pair(0x17, "or"),
+  Pair(0x18, "xor"),
+  Pair(0x19, "not"),
+  Pair(0x1a, "byte"),
+  Pair(0x20, "sha3"),
+  Pair(0x30, "address"),
+  Pair(0x31, "balance"),
+  Pair(0x32, "origin"),
+  Pair(0x33, "caller"),
+  Pair(0x34, "callvalue"),
+  Pair(0x35, "calldataload"),
+  Pair(0x36, "calldatasize"),
+  Pair(0x37, "calldatacopy"),
+  Pair(0x38, "codesize"),
+  Pair(0x39, "codecopy"),
+  Pair(0x3a, "gasPrice"),
+  Pair(0x3b, "extcodesize"),
+  Pair(0x40, "blockhash"),
+  Pair(0x41, "coinbase"),
+  Pair(0x42, "timestamp"),
+  Pair(0x43, "number"),
+  Pair(0x44, "difficulty"),
+  Pair(0x45, "gaslimit"),
+  Pair(0x50, "pop"),
+  Pair(0x51, "mload"),
+  Pair(0x52, "mstore"),
+  Pair(0x53, "mstore8"),
+  Pair(0x54, "sload"),
+  Pair(0x55, "sstore"),
+  Pair(0x56, "jump"),
+  Pair(0x57, "jumpi"),
+  Pair(0x58, "pc"),
+  Pair(0x59, "msize"),
+  Pair(0x5a, "gas"),
+  Pair(0x5b, "jumpdest"),
+  Pair(0x60, "push1"),
+  Pair(0x61, "push2"),
+  Pair(0x62, "push3"),
+  Pair(0x63, "push4"),
+  Pair(0x64, "push5"),
+  Pair(0x65, "push6"),
+  Pair(0x66, "push7"),
+  Pair(0x67, "push8"),
+  Pair(0x68, "push9"),
+  Pair(0x69, "push10"),
+  Pair(0x6a, "push11"),
+  Pair(0x6b, "push12"),
+  Pair(0x6c, "push13"),
+  Pair(0x6d, "push14"),
+  Pair(0x6e, "push15"),
+  Pair(0x6f, "push16"),
+  Pair(0x70, "push17"),
+  Pair(0x71, "push18"),
+  Pair(0x72, "push19"),
+  Pair(0x73, "push20"),
+  Pair(0x74, "push21"),
+  Pair(0x75, "push22"),
+  Pair(0x76, "push23"),
+  Pair(0x77, "push24"),
+  Pair(0x78, "push25"),
+  Pair(0x79, "push26"),
+  Pair(0x7a, "push27"),
+  Pair(0x7b, "push28"),
+  Pair(0x7c, "push29"),
+  Pair(0x7d, "push30"),
+  Pair(0x7e, "push31"),
+  Pair(0x7f, "push32"),
+  Pair(0x90.toByte(), "swap1"),
+  Pair(0x91.toByte(), "swap2"),
+  Pair(0x92.toByte(), "swap3"),
+  Pair(0x93.toByte(), "swap4"),
+  Pair(0x94.toByte(), "swap5"),
+  Pair(0x95.toByte(), "swap6"),
+  Pair(0x96.toByte(), "swap7"),
+  Pair(0x97.toByte(), "swap8"),
+  Pair(0x98.toByte(), "swap9"),
+  Pair(0x99.toByte(), "swap10"),
+  Pair(0x9a.toByte(), "swap11"),
+  Pair(0x9b.toByte(), "swap12"),
+  Pair(0x9c.toByte(), "swap13"),
+  Pair(0x9d.toByte(), "swap14"),
+  Pair(0x9e.toByte(), "swap15"),
+  Pair(0x9f.toByte(), "swap16"),
+  Pair(0xf3.toByte(), "return"),
+  Pair(0xa0.toByte(), "log0"),
+  Pair(0xa1.toByte(), "log1"),
+  Pair(0xa2.toByte(), "log2"),
+  Pair(0xa3.toByte(), "log3"),
+  Pair(0xa4.toByte(), "log4"),
+  Pair(0xfe.toByte(), "invalid"),
+  Pair(0xff.toByte(), "selfdestruct"),
+)
